Refactor project structure; add initial tests;
