+++ .gitignore
@@ -6,3 +6,6 @@
 
 # LaTex
 *tex
+
+# Configuration
+*config.yml
