+++ monitorframe/datamodel.py
@@ -1,0 +1,136 @@
+import abc
+import numpy as np
+import pandas as pd
+import peewee
+
+from itertools import repeat
+from playhouse.reflection import generate_models
+from typing import List, Dict, Union
+
+from .database import DATA_DB
+
+
+class DataInterface(abc.ABC):
+
+    @abc.abstractmethod
+    def get_new_data(self) -> Union[List[dict], Dict[str, list]]:
+        pass
+
+    @abc.abstractmethod
+    def ingest(self):
+        pass
+
+
+class PandasMeta(abc.ABCMeta):
+    def __new__(mcs, classnames, bases, class_dict):
+        class_dict['get_new_data'] = mcs.wrap(class_dict['get_new_data'])
+
+        return super(PandasMeta, mcs).__new__(mcs, classnames, bases, class_dict)
+
+    @staticmethod
+    def wrap(get_new_data):
+        def to_pandas(self):
+            data = get_new_data(self)
+            df = pd.DataFrame(data)
+
+            return df
+
+        return to_pandas
+
+
+# noinspection PyAbstractClass
+# Partial implementation
+class BaseDataModel(DataInterface, metaclass=PandasMeta):
+    """Baseclass for monitor data models.
+
+    Intended to be subclassed with one required method: get_data. Results from get_data will be used to generate a
+    pandas DataFrame which the monitors use for the data source.
+    """
+    _database = DATA_DB
+
+    def __init__(self, find_new=True):
+        self._array_types = None
+        self.new_data = None
+        self._formatted_data = None
+
+        self.table_name = self.__class__.__name__
+
+        # Read in and ingest new data
+        if find_new:
+            self.new_data = self.get_new_data()
+            self._array_types = self._find_array_types()
+            self._formatted_data = self._format_for_ingest()
+
+    # noinspection PyUnresolvedReferences
+    # noinspection PyCompatibility
+    # self.new_data will be a pandas DataFrame object
+    def _find_array_types(self):
+        """Find datatypes in the dataframe that are likely arrays of some kind."""
+        supported = [list, np.ndarray]  # Supported array types
+
+        example = self.new_data.iloc[0]  # All rows should be the same.. otherwise ingestion won't even get this far
+
+        # Assuming that "object" types that aren't strings are arrays
+        return [
+            key for key, dtype in self.new_data.dtypes.iteritems() if dtype == 'O' and type(example[key]) in supported
+        ]
+
+    # noinspection PyUnresolvedReferences
+    # self.new_data will be a pandas DataFrame object
+    def _format_for_ingest(self):
+        """Format new data for ingest. Primarily, if there are arrays as elements in any column, convert those to
+        strings.
+        """
+        if self._array_types:  # If there are array elements, convert to string. Else, do nothing
+            ingestible = self.new_data.copy()
+
+            for key in self._array_types:
+                ingestible[key] = ingestible[key].astype(str)
+
+            return ingestible
+
+        return self.new_data
+
+    @property
+    def model(self):
+        """Return the database table model object if the table exists in the database."""
+        try:
+            return generate_models(
+                self._database, literal_column_names=True, table_names=[self.table_name]
+            )[self.table_name]
+
+        except KeyError:
+            return
+
+    @abc.abstractmethod
+    def get_new_data(self) -> Union[List[dict], Dict[str, list]]:
+        """Retrieve monitor data. Should return row-wise or column-wise data."""
+        pass
+
+    # noinspection PyUnresolvedReferences
+    # self._formatted_data will be a pandas DataFrame object
+    def ingest(self):
+        """Ingest new data into database."""
+        with self._database as db:
+            self._formatted_data.to_sql(self.table_name, db, if_exists='append', index=False)
+
+    def query_to_pandas(self, query: peewee.ModelSelect, array_cols: list = None, array_dtypes: list = None
+                        ) -> pd.DataFrame:
+        """Convert a model query to a pandas dataframe."""
+        df = pd.DataFrame(query.dicts())
+
+        if not array_cols:
+            array_cols = self._array_types  # Try to use the new data to infer what the format should be
+
+        if array_cols:
+            if not array_dtypes:
+                array_dtypes = repeat(float, len(array_cols))
+
+            # Convert array columns to numpy arrays. Assume the dtype should be a float if not specified
+            for key, dtype in zip(array_cols, array_dtypes):
+                df[key] = df[key].apply(
+                    lambda x: np.array(x.strip('[]').split(', '), dtype=float) if ',' in x
+                    else np.array(x.strip('[]').split(), dtype=float)
+                )
+
+        return df
